requirements.txt
fastapi
uvicorn
pydantic
requests
sqlalchemy
pillow
opencv-python
google-cloud-vision
python-multipart
beautifulsoup4
